# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AMTS Pro – Multi-Strategy Trading Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SignalType = (typeof SignalType)[keyof typeof SignalType];

export const SignalType = {
  BUY: "BUY",
  SELL: "SELL",
  WAIT: "WAIT",
} as const;

export type StrategyMode = (typeof StrategyMode)[keyof typeof StrategyMode];

export const StrategyMode = {
  BEGINNER: "BEGINNER",
  PRO: "PRO",
} as const;

export type StrategyCategory =
  (typeof StrategyCategory)[keyof typeof StrategyCategory];

export const StrategyCategory = {
  TREND: "TREND",
  RANGE: "RANGE",
  MOMENTUM: "MOMENTUM",
  BREAKOUT: "BREAKOUT",
  SCALPING: "SCALPING",
  SWING: "SWING",
  PRICE_ACTION: "PRICE_ACTION",
  SMART_MONEY: "SMART_MONEY",
  SUPPLY_DEMAND: "SUPPLY_DEMAND",
  INDICATOR: "INDICATOR",
  QUANT: "QUANT",
  CRYPTO: "CRYPTO",
  HIGH_RISK: "HIGH_RISK",
} as const;

export type RiskLevel = (typeof RiskLevel)[keyof typeof RiskLevel];

export const RiskLevel = {
  LOW: "LOW",
  MEDIUM: "MEDIUM",
  HIGH: "HIGH",
} as const;

export interface Pair {
  symbol: string;
  label: string;
}

export interface Timeframe {
  value: string;
  label: string;
  description: string;
}

export interface Strategy {
  key: string;
  name: string;
  category: StrategyCategory;
  mode: StrategyMode;
  riskLevel: RiskLevel;
  description: string;
  logic?: string;
}

export interface Candle {
  openTime: number;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}

export interface Ticker {
  pair: string;
  lastPrice: number;
  change24h: number;
  changePercent24h: number;
  high24h: number;
  low24h: number;
  volume24h: number;
  updatedAt: string;
}

export interface Signal {
  id: string;
  strategyKey: string;
  strategyName: string;
  category: StrategyCategory;
  signal: SignalType;
  /**
   * @minimum 0
   * @maximum 100
   */
  confidence: number;
  entry: number;
  stopLoss: number;
  takeProfit: number;
  pair: string;
  timeframe: string;
  rationale: string;
  riskLevel: RiskLevel;
  createdAt: string;
}

export interface TopSignalResponse {
  top: Signal | null;
  supporting: Signal[];
  conflicting: Signal[];
  totalStrategies: number;
}

export interface ConflictReport {
  hasConflict: boolean;
  buyCount: number;
  sellCount: number;
  waitCount: number;
  dominant: SignalType;
  message: string;
}

export interface StrategyPerformance {
  strategyKey: string;
  strategyName: string;
  totalSignals: number;
  buyCount: number;
  sellCount: number;
  waitCount: number;
  avgConfidence: number;
}

export interface SignalDistribution {
  buy: number;
  sell: number;
  wait: number;
}

export interface CategoryBreakdown {
  category: StrategyCategory;
  signal: SignalType;
  avgConfidence: number;
  count: number;
}

export interface DashboardSummary {
  pair: string;
  timeframe: string;
  ticker: Ticker;
  top: Signal | null;
  distribution: SignalDistribution;
  categoryBreakdown: CategoryBreakdown[];
  conflict: ConflictReport;
  lastUpdated: string;
  strategiesActive: number;
}

export type GetStrategyPerformanceParams = {
  pair: string;
  timeframe: string;
};

export type GetTickerParams = {
  pair: string;
};

export type GetCandlesParams = {
  pair: string;
  timeframe: string;
  /**
   * @minimum 10
   * @maximum 500
   */
  limit?: number;
};

export type GetTopSignalParams = {
  pair: string;
  timeframe: string;
};

export type GetAllSignalsParams = {
  pair: string;
  timeframe: string;
};

export type GetConflictsParams = {
  pair: string;
  timeframe: string;
};

export type GetSignalHistoryParams = {
  pair: string;
  timeframe: string;
  /**
   * @minimum 1
   * @maximum 500
   */
  limit?: number;
  strategyKey?: string;
};

export type GetDashboardSummaryParams = {
  pair: string;
  timeframe: string;
};