# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AMTS Pro – Multi-Strategy Trading Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryBreakdown } from "./categoryBreakdown";
import type { ConflictReport } from "./conflictReport";
import type { Signal } from "./signal";
import type { SignalDistribution } from "./signalDistribution";
import type { Ticker } from "./ticker";

export interface DashboardSummary {
  pair: string;
  timeframe: string;
  ticker: Ticker;
  top: Signal | null;
  distribution: SignalDistribution;
  categoryBreakdown: CategoryBreakdown[];
  conflict: ConflictReport;
  lastUpdated: Date;
  strategiesActive: number;
}