# lib/api-zod/src/generated/types/signalType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AMTS Pro – Multi-Strategy Trading Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export type SignalType = (typeof SignalType)[keyof typeof SignalType];

export const SignalType = {
  BUY: "BUY",
  SELL: "SELL",
  WAIT: "WAIT",
} as const;