# lib/api-zod/src/generated/types/strategyCategory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AMTS Pro – Multi-Strategy Trading Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export type StrategyCategory =
  (typeof StrategyCategory)[keyof typeof StrategyCategory];

export const StrategyCategory = {
  TREND: "TREND",
  RANGE: "RANGE",
  MOMENTUM: "MOMENTUM",
  BREAKOUT: "BREAKOUT",
  SCALPING: "SCALPING",
  SWING: "SWING",
  PRICE_ACTION: "PRICE_ACTION",
  SMART_MONEY: "SMART_MONEY",
  SUPPLY_DEMAND: "SUPPLY_DEMAND",
  INDICATOR: "INDICATOR",
  QUANT: "QUANT",
  CRYPTO: "CRYPTO",
  HIGH_RISK: "HIGH_RISK",
} as const;