# lib/api-zod/src/generated/types/timeframe.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AMTS Pro – Multi-Strategy Trading Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Timeframe {
  value: string;
  label: string;
  description: string;
}